Created hook folder, placed useTheme in here. Refactored eventsPage by extracting hook logic to their own seperate files in the hook folder.
