+++ src/components/layouts/EventPage.tsx
@@ -1,46 +1,13 @@
-import { useEffect, useState } from 'react';
 import NavBar from '../major/nav-bar';
 import SideBar from '../major/side-bar';
-import { FiltersState } from '../major/side-bar/types.ts';
 import EventsContainer from '../major/eventsContainer.tsx';
-import { getEventByMode } from '../../api/services/events.ts';
-import { filterEvents } from '../../utils/filterEvents.ts';
-import { Event } from 'models/event.model.ts';
+import { useEventFilters } from '../../hooks/useEventFilters.ts';
+import { useEvents } from '../../hooks/useEvents.ts';
 
+// Create hook folder and extract the filter and event filter into it. along with usetheme.
 export default function EventLayout() {
-  const [filters, setFilters] = useState<FiltersState>({
-    selectedModes: ['music'],
-    search: '',
-    price: 50,
-    distance: 20,
-    date: 'this month',
-  });
-
-  const updateFilters = (updates: Partial<FiltersState>) => {
-    setFilters((prev) => ({
-      ...prev,
-      ...updates,
-    }));
-  };
-
-  const [rawEvents, setRawEvents] = useState<Event[]>([]);
-  const [filteredEvents, setFilteredEvents] = useState<Event[]>([]);
-
-  useEffect(() => {
-    const getEvents = async () => {
-      try {
-        const data = await getEventByMode(filters.selectedModes);
-        setRawEvents(data);
-      } catch (error) {
-        new Error('Error setting events by mode: ' + error);
-      }
-    };
-    getEvents();
-  }, [filters.selectedModes]);
-
-  useEffect(() => {
-    setFilteredEvents(filterEvents(rawEvents, filters));
-  }, [rawEvents, filters]);
+  const { filters, updateFilters } = useEventFilters();
+  const { filteredEvents } = useEvents(filters);
 
   return (
     <div className="flex flex-col min-h-screen">
